+++ apple-codesign/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 Released on ReleaseDate.
 
+## 0.22.0
+
+Released on 2022-12-21.
+
 * Cargo.toml now defines patch version for all dependencies.
 * goblin crate upgraded from 0.5 to 0.6.
 * App Store Connect API code extracted to its own crate, `app-store-connect`.